#3843 separate ruleset table & match formatting + service fix
