+++ src/frontend/src/apis/rules.api.ts
@@ -34,4 +34,3 @@ export const getRulesetsByRulesetType = (rulesetTypeId: string) => {
     transformResponse: (data) => JSON.parse(data)
   });
 };
-
